refactor(runtime): split methods_distribution.rs into cohesive modules (§7-8)
`methods_distribution.rs` was 1209 lines, well over the 500-line module limit.
Split it along its natural seams into four files, each under 360 lines, with no
behavior change:

- `methods_distribution.rs` (343) — CompUnit/Distribution method dispatch entry
  points (`dispatch_distribution_method`, `dispatch_cur_installation_method`),
  the globalish-package helpers, and the `run-script` repository loader.
- `methods_distribution_cur_resolve.rs` (341) — candidate *resolution*: file
  hash building, depspec extraction/matching, FileSystem/Installation candidate
  enumeration.
- `methods_distribution_cur_inst.rs` (353) — Installation mutations: install,
  uninstall, list installed, and resolve+load a `need`ed dependency.
- `methods_distribution_helpers.rs` (211) — the dependency-free JSON
  parser/serializer plus small utilities (`dist_identity`,
  `platform_library_name`, `hash_strings`), made `pub(super)` so the sibling
  `impl Interpreter` modules can call them.

The private `matches_depspec` keeps its visibility since both callers
(`cur_fs_candidates`, `cur_inst_candidates`) stay together in `_cur_resolve`.
First slice of the §7-8 file-splitting sweep (500-line rule).
