+++ src/runtime/mod.rs
@@ -166,6 +166,9 @@ mod methods_dispatch_match2;
 mod methods_dispatch_match3;
 mod methods_dispatch_new;
 mod methods_distribution;
+mod methods_distribution_cur_inst;
+mod methods_distribution_cur_resolve;
+mod methods_distribution_helpers;
 mod methods_enum_dispatch;
 mod methods_format;
 mod methods_grammar;
